refactor: move ping_and_wait out of provider.rs
