+++ src/dfx/src/lib/provider.rs
@@ -16,15 +16,12 @@ use dfx_core::error::network_config::NetworkConfigError::{
 };
 use dfx_core::identity::WALLET_CONFIG_FILENAME;
 
-use anyhow::{anyhow, bail, Context};
+use anyhow::anyhow;
 use fn_error_context::context;
-use ic_agent::agent::http_transport::ReqwestHttpReplicaV2Transport;
-use ic_agent::Agent;
 use lazy_static::lazy_static;
 use slog::{debug, info, warn, Logger};
 use std::path::{Path, PathBuf};
 use std::sync::{Arc, RwLock};
-use std::time::Duration;
 use url::Url;
 
 lazy_static! {
@@ -406,40 +403,6 @@ pub fn parse_provider_url(url: &str) -> Result<String, NetworkConfigError> {
     Url::parse(url)
         .map(|_| String::from(url))
         .map_err(|e| ParseProviderUrlFailed(Box::new(url.to_string()), e))
-}
-
-pub async fn ping_and_wait(url: &str) -> DfxResult {
-    let agent = Agent::builder()
-        .with_transport(
-            ReqwestHttpReplicaV2Transport::create(url)
-                .with_context(|| format!("Failed to create replica transport from url {url}.",))?,
-        )
-        .build()
-        .with_context(|| format!("Failed to build agent with url {url}."))?;
-    let mut retries = 0;
-    loop {
-        let status = agent.status().await;
-        match status {
-            Ok(status) => {
-                let healthy = match &status.replica_health_status {
-                    Some(status) if status == "healthy" => true,
-                    None => true, // emulator doesn't report replica_health_status
-                    _ => false,
-                };
-                if healthy {
-                    break;
-                }
-            }
-            Err(e) => {
-                if retries >= 60 {
-                    bail!(e);
-                }
-                tokio::time::sleep(Duration::from_secs(1)).await;
-                retries += 1;
-            }
-        }
-    }
-    Ok(())
 }
 
 #[cfg(test)]
